requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
mysqlclient
alembic
passlib
bcrypt==4.0.1
pyjwt
python-dotenv
